ND2: catch potential long overflow
d79a091 should have included this.

+++ components/bio-formats/src/loci/formats/in/NativeND2Reader.java
@@ -1485,12 +1485,12 @@ public class NativeND2Reader extends FormatReader {
             value = in.readCString();
             break;
           case (9): // ByteArray
-            int length = (int) in.readLong();
+            long length = in.readLong();
             if (length + in.getFilePointer() > stop) {
               in.seek(stop);
               continue;
             }
-            byte[] data = new byte[length];
+            byte[] data = new byte[(int) length];
             in.read(data);
             value = java.util.Arrays.toString(data); // todo
             break;
